Check in - partially working. CLI entrypoint needs work (see notes below)
Flow now works as follows:

user creates an `app` via cli command. This can either be done via a loadout file (working),
or via cli switches (needs fixing - hasn't been updated)

This allows the runtime config, polling-interval etc, to be split out as part of the keeper `run`
command.

User then populates `app` directories with whatever files they want. This can be as granular or relaxed
as the user likes. Introduces the concept of groups, there is a base group called `all` which every
component is a part of. Any files in this group will get transferred to agents, (depending on any state
directives) components then opt-in to other groups; if they do so, they will receives this groups files / dirs.

There is also a `staging` area for each component where files / dirs specific to that component can be added.

At runtime, these groups and staging objects get symlinked into a `fake_root` directory - this fake_root dir is
essentially / on the remote machine. If the user wants, they can skip the groups and staging, and work directly with
the fake_root.

Upon app termination, all dirs / files that have been linked, get removed. The created links are
stored in a file, so if the app crashes or the power is unplugged or something, the app can recover on boot.

State directives - these are used to determine file state on the remote machine. They serve 2 purposes - to either
direct where the files / dirs go that are in groups or staging, or to set a sync state for the file or directory

Sync states are one of the following (enum): ENSURE_CREATED (default), ALLOW_ADDS, or STRICT

ENSURE_CREATED, as long as the path exists on the remote machine, nothing is changed
ALLOW_ADDS, applies to dirs, objects are allowed to be added, but cannot be removed
STRICT - CRC must match exactly. If not, entire object gets replaced.

Note on crc - this has been updated now to include directory names, this all happens recursively.

If no state directive exists in the config, any files / dirs are copied to the agents working_dir (as specified in the config)

Config inheritance - config can be specified on the app level or the component level, if not specified on the component
level, it is inherited from the app level, then uses global defaults. E.g. the working_dir by default is /tmp